read: update default main branch to prod
Because I found master is an accepted main branch for long time, but
it's not a real MASTER. I have no idea why it's called MASTER, there are
no slaves!

So I think prod, staging, dev ... is a good idea. It make name is clear
and understandable as what will code in this branch are used.